requirements.txt

###### Requirements without Version Specifiers ######
matplotlib
sklearn
scipy
tqdm

###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
numpy == 1.16.0
pandas == 0.25.3
GPy == 1.9.9
munch == 2.3.2
opencv-contrib-python==4.1.0.25
tensorflow-gpu == 1.14             # Version Matching. Must be version 0.6.1
tensorflow_probability == 0.7.0            # Minimum version 0.8.0


